requirements.txt
numpy>=1.20.0
pandas>=1.3.0
torch>=1.10.0
pytorch-lightning>=1.5.0
luigi>=3.0.0
wandb>=0.12.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=1.0.0
zarr>=2.8.0
pyyaml>=6.0
tqdm>=4.62.0
